update Client, Item add Order

+++ src/main/java/com/example/springlessontest/model/dao/item/dbDaoItem.java
@@ -38,9 +38,11 @@ public class dbDaoItem implements IDaoItem{
 
     @Override
     public Item delete(Integer id) {
-        Item item = new Item(id, itemRepository.findById(id).get().getItemName(),
-                itemRepository.findById(id).get().getItemArticle());
-        itemRepository.deleteById(id);
-        return item;
+        if(itemRepository.findById(id).isPresent()) {
+            Item item = itemRepository.findById(id).get();
+            itemRepository.deleteById(id);
+            return item;
+        }
+        return null;
     }
 }
